audio_stego/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.26.3
scipy>=1.15.0
pydub>=0.25.1
librosa>=0.10.1
soundfile>=0.12.1
matplotlib>=3.8.3
Pillow>=10.2.0
requests>=2.31.0
werkzeug>=3.0.1
PyWavelets>=1.6.0
llama-cpp-python>=0.2.50
